separate driver and trainer
