--- AdvectionDiffusion/Coupling/PINN/pinnschwarz/README.md
@@ -1,6 +1,0 @@
-# Add yaml file
-Please install the yaml library:
-```python
-pip3 install pyyaml
-```
-
