+++ AdvectionDiffusion/Coupling/PINN/pyproject.toml
@@ -14,6 +14,7 @@ dependencies = [
     "numpy",
     "matplotlib",
     "pandas",
+    "ray[tune]",
     "pyyaml",
     "black",
 ]
